+++ simple_graphql/django/config.py
@@ -1,5 +1,5 @@
 from dataclasses import fields
-from typing import Optional, Union, cast
+from typing import Optional, Type, Union, cast
 
 from simple_graphql.django.types import (
     ModelClass,
@@ -19,6 +19,13 @@ def extract_schema_config(config: Optional[ModelConfig]) -> Optional[ModelSchema
             field.name: getattr(config, field.name, None)
             for field in fields(ModelSchemaConfig)
         }
+    )
+
+
+def extract_extra_meta_config(extra_meta: Optional[Type]) -> ModelSchemaConfig:
+    return (
+        extract_schema_config(getattr(extra_meta, "config", None))
+        or ModelSchemaConfig.get_defaults()
     )