+++ example/schema.py
@@ -35,4 +35,38 @@ class GetUserInfoMutation(relay.ClientIDMutation):
         )
 
 
+@schema.graphql_query("math")
+class Math(graphene.ObjectType):
+    add = graphene.Field(
+        graphene.NonNull(graphene.Int),
+        a=graphene.NonNull(graphene.Int),
+        b=graphene.NonNull(graphene.Int),
+    )
+    substract = graphene.Field(
+        graphene.NonNull(graphene.Int),
+        a=graphene.NonNull(graphene.Int),
+        b=graphene.NonNull(graphene.Int),
+    )
+
+    def resolve_add(self, info: Any, a: int, b: int, **kwargs: Any) -> int:
+        return a + b
+
+    def resolve_substract(self, info: Any, a: int, b: int, **kwargs: Any) -> int:
+        return a - b
+
+
+class Echo(graphene.ObjectType):
+    ping = graphene.Field(
+        graphene.NonNull(graphene.String), input=graphene.NonNull(graphene.String)
+    )
+
+    def resolve_ping(self, info: Any, input: str, **kwargs: Any) -> str:
+        if input == "ping":
+            return "pong"
+        return input
+
+
 schema.register_mutation("login", LoginMutation.Field())
+schema.register_query(
+    "echo", graphene.Field(Echo, required=True), lambda *args, **kwargs: Echo()
+)