+++ simple_graphql/django/schema/exceptions.py
@@ -12,7 +12,14 @@ class ModelAlreadyRegistered(Exception):
 class MutationAlreadyRegistered(Exception):
     def __init__(self, name: str):
         super().__init__(
-            f"Mutation {name} " "has already been registered to the GraphQL schema"
+            f"Mutation {name} has already been registered to the GraphQL schema"
+        )
+
+
+class QueryAlreadyRegistered(Exception):
+    def __init__(self, name: str):
+        super().__init__(
+            f"Query {name} has already been registered to the GraphQL schema"
         )