+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-- Nothing
+- Added Schema.register_query method for registering queries
+- Added Schema.graphql_query decorator for registering queries
+- Improved logging of schema building errors
 
 ## 0.3.0